QA 버그 픽스 (#65)
* feat: 밤 결과 조회 상태 NOTICE 로 변경

* feat: 마피아 안죽이기 기능 구현

* feat: 투표하기 투표 인원 집계시 죽은사람 제외

* feat: 방 재시작 이후 채팅 초기화

+++ src/main/java/mafia/mafiatogether/domain/Vote.java
@@ -62,7 +62,7 @@ public class Vote {
         playerVote.clear();
     }
 
-    public boolean isAllParticipatedVote(final int total) {
+    public boolean isAllParticipatedVote(final Long total) {
         return this.playerVote.keySet().size() == total;
     }
 }
